backend/fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
ortools>=9.8.3296
pulp==2.7.0
numpy>=1.26.0
pandas>=2.1.3
python-decouple==3.8
pytest==7.4.3
httpx==0.25.2
pydantic-settings==2.1.0